Complete aruco tag integration (for LPSC)
This commit contains several changes to the workflow and structure of
the operation of the camera. The Aruco tags are squirrely and our 8'x2' field
was at the boundary of what we could do with the gopro 11 camera from one side.
Several todo's show configurable variables that should be moved to the yaml.